Moved around some code
Sending the announcement message now has priority over removing the old one. Issue #7 was not an issue with the code but with the files locally on the bot. Fix hasn't been tested yet.

Also moved the game.concluded=true statement in the `!game stop` command down in case of errors, and replaced some console.logs to console.errors and console.warns.